chore: ajusting the comands

+++ scripts/installing/teste.sh
@@ -3,9 +3,9 @@ GREEN='\033[0;32m'
 YELLOW='\033[1;33m'
 NC='\033[0m'
 
-INSTALL_DIR="$HOME/Tools"
-ALIAS_FILE="$INSTALL_DIR/tools.bash"
-BASHRC="$HOME/.bashrc"
+INSTALL_DIR="$HOME/tools"
+ALIAS_FILE="$INSTALL_DIR/tools.sh"
+BASHRC="$HOME/.bash_config/aliases.sh"
 BLACK_WHOLE="/dev/null"
 
 echo -e "${YELLOW}Iniciando configuração modular...${NC}"
@@ -45,25 +45,30 @@ echo "----------------------------------------------------------------"
 # ==============================================================================
 # GODOT ENGINE (Criação de Jogos 2D/3D - Leve e Poderoso)
 # ==============================================================================
-echo -e "${YELLOW}[Extra 7] Godot Engine...${NC}"
+echo -e "${YELLOW}Godot Engine...${NC}"
 
-cd "$INSTALL_DIR"
-rm -rf godot-dir
-mkdir -p godot-dir
-cd godot-dir
+if [ "$IS_ADMIN" = true ]; then
+    #versão meio pesada
+    sudo snap install godot4
+else
+    cd "$INSTALL_DIR"
+    rm -rf godot-dir
+    mkdir -p godot-dir
+    cd godot-dir
 
-echo "Baixando Godot 4.3 (Stable)..."
-# O Godot vem num ZIP simples, não é AppImage
-wget -c "https://github.com/godotengine/godot/releases/download/4.3-stable/Godot_v4.3-stable_linux.x86_64.zip" -O godot.zip
+    echo "Baixando Godot 4.3 (Stable)..."
+    # O Godot vem num ZIP simples, não é AppImage
+    wget -c "https://github.com/godotengine/godot/releases/download/4.3-stable/Godot_v4.3-stable_linux.x86_64.zip" -O godot.zip
 
-echo "Extraindo..."
-unzip godot.zip > /dev/null
-rm godot.zip
-# O executável tem um nome longo, vamos simplificar no alias
+    echo "Extraindo..."
+    unzip godot.zip > /dev/null
+    rm godot.zip
+    mv Godot_v4.3-stable_linux.x86_64 godot4_
 
-if ! grep -q "alias godot=" "$ALIAS_FILE"; then
-    # Godot precisa de GPU. Se der erro no lab, tente adicionar --headless (mas aí não tem interface)
-    echo "alias godot=\"$INSTALL_DIR/godot-dir/Godot_v4.3-stable_linux.x86_64 > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
+    if ! grep -q "alias godot=" "$ALIAS_FILE"; then
+        # Godot precisa de GPU. Se der erro no lab, tente adicionar --headless (mas aí não tem interface)
+        echo "alias godot=\"$INSTALL_DIR/godot-dir/godot4_ > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
+    fi
 fi
 
 echo "----------------------------------------------------------------"
@@ -72,7 +77,7 @@ echo "----------------------------------------------------------------"
 # KICAD (Design de Circuitos e PCBs)
 # Nota: KiCad é complexo. Usamos um AppImage não-oficial (SSB) pois o oficial exige root.
 # ==============================================================================
-echo -e "${YELLOW}[Extra 8] KiCad (PCB Design)...${NC}"
+echo -e "${YELLOW}KiCad (PCB Design)...${NC}"
 
 cd "$INSTALL_DIR"
 rm -rf kicad-dir
@@ -81,13 +86,28 @@ cd kicad-dir
 
 echo "Baixando KiCad 7 (AppImage com bibliotecas inclusas)..."
 # Usamos a versão 7 pois é a mais estável em formato AppImage atualmente
-wget -c "https://github.com/probonopd/KiCad/releases/download/continuous/KiCad-7.0.0-202302130704+6a380965f7-x86_64.AppImage" -O kicad.AppImage
+wget -c "https://github.com/KarlZeilhofer/kicad-appimage/releases/download/v5.1.4/KiCad-5.1.4.glibc2.27-x86_64.AppImage" -O kicad.AppImage
 chmod +x kicad.AppImage
 
 echo "Extraindo KiCad (Isso pode demorar, é grande)..."
 ./kicad.AppImage --appimage-extract > /dev/null
 mv squashfs-root kicad-files
 rm kicad.AppImage
+
+# removendo algumas coisas
+if [ "$IS_ADMIN" = true ]; then
+    find . -name "libpixman-1.so*" -delete
+    find . -name "libcairo.so*" -delete
+    find . -name "libstdc++.so.6" -delete
+
+    find . -name "libglib-2.0.so*" -delete
+    find . -name "libgmodule-2.0.so*" -delete
+    find . -name "libgio-2.0.so*" -delete
+    find . -name "libgobject-2.0.so*" -delete
+    
+    find . -name "libselinux.so*" -delete
+fi 
+
 
 if ! grep -q "alias kicad=" "$ALIAS_FILE"; then
     echo "alias kicad=\"$INSTALL_DIR/kicad-dir/kicad-files/AppRun > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
@@ -98,34 +118,38 @@ echo "----------------------------------------------------------------"
 # ==============================================================================
 # CIRCUITBLOCKS (Programação Visual de Hardware)
 # ==============================================================================
-echo -e "${YELLOW}[Extra 9] CircuitBlocks...${NC}"
+echo -e "${YELLOW}CircuitBlocks...${NC}"
 
-cd "$INSTALL_DIR"
-rm -rf circuitblocks-dir
-mkdir -p circuitblocks-dir
-cd circuitblocks-dir
+if [ "$IS_ADMIN" = true ]; then
+    cd "$INSTALL_DIR"
+    rm -rf circuitblocks-dir
+    mkdir -p circuitblocks-dir
+    cd circuitblocks-dir
 
-echo "Baixando CircuitBlocks..."
-# Link da versão 1.10.1
-wget -c "https://github.com/CircuitMess/CircuitBlocks/releases/download/v1.10.1/circuitblocks-1.10.1-x86_64.AppImage" -O cb.AppImage
-chmod +x cb.AppImage
+    echo "Baixando CircuitBlocks..."
+    # Link da versão 1.10.1
+    wget -c "https://github.com/CircuitMess/CircuitBlocks/releases/download/v1.10.0/CircuitBlocks-1.10.0-Linux.AppImage" -O cb.AppImage
+    chmod +x cb.AppImage
 
-echo "Extraindo..."
-./cb.AppImage --appimage-extract > /dev/null
-mv squashfs-root cb-files
-rm cb.AppImage
+    echo "Extraindo..."
+    ./cb.AppImage --appimage-extract > /dev/null
+    mv squashfs-root cb-files
+    rm cb.AppImage
 
-if ! grep -q "alias circuitblocks=" "$ALIAS_FILE"; then
-    # É Electron, precisa das flags de segurança
-    echo "alias circuitblocks=\"$INSTALL_DIR/circuitblocks-dir/cb-files/circuitblocks --no-sandbox --disable-gpu > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
-fi
+    if ! grep -q "alias circuitblocks=" "$ALIAS_FILE"; then
+        # É Electron, precisa das flags de segurança
+        echo "alias circuitblocks=\"$INSTALL_DIR/circuitblocks-dir/cb-files/circuitblocks --no-sandbox --disable-gpu > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
+    fi
+fi else
+    echo "CircuitBlocks não pôde ser instalado\n"
+fi 
 
 echo "----------------------------------------------------------------"
 
 # ==============================================================================
 # MUSIC BLOCKS (Programação Visual de Música)
 # ==============================================================================
-echo -e "${YELLOW}[Extra 10] Music Blocks...${NC}"
+echo -e "${YELLOW}Music Blocks...${NC}"
 
 cd "$INSTALL_DIR"
 rm -rf musicblocks-dir
@@ -134,16 +158,14 @@ cd musicblocks-dir
 
 echo "Baixando Music Blocks Launcher..."
 # Versão 2.1.2
-wget -c "https://github.com/sugarlabs/musicblocks-launcher/releases/download/v2.1.2/Music.Blocks.Launcher-2.1.2.AppImage" -O mb.AppImage
-chmod +x mb.AppImage
+wget -c "https://github.com/sugarlabs/musicblocks/archive/refs/tags/v3.6.2.tar.gz" -O mb.tar.gz
 
 echo "Extraindo..."
-./mb.AppImage --appimage-extract > /dev/null
-mv squashfs-root mb-files
-rm mb.AppImage
+tar -xvf mb.tar.gz
+rm mb.tar.gz
 
 if ! grep -q "alias musicblocks=" "$ALIAS_FILE"; then
-    echo "alias musicblocks=\"$INSTALL_DIR/musicblocks-dir/mb-files/AppRun --no-sandbox --disable-gpu > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
+    echo "alias musicblocks=\"$INSTALL_DIR/musicblocks-dir/mb-bin/musicblocks --no-sandbox --disable-gpu > /dev/null 2>&1 &\"" >> "$ALIAS_FILE"
 fi
 
 echo "----------------------------------------------------------------"
@@ -151,7 +173,7 @@ echo "----------------------------------------------------------------"
 # ==============================================================================
 # ARCADE MANAGER (Gerenciador de ROMs/Retropie)
 # ==============================================================================
-echo -e "${YELLOW}[Extra 11] Arcade Manager...${NC}"
+echo -e "${YELLOW}Arcade Manager...${NC}"
 
 cd "$INSTALL_DIR"
 rm -rf arcademanager-dir
@@ -160,7 +182,7 @@ cd arcademanager-dir
 
 echo "Baixando Arcade Manager..."
 # Versão 7.1
-wget -c "https://github.com/cosmo0/arcade-manager/releases/download/v7.1/arcade-manager-7.1.0-linux.AppImage" -O am.AppImage
+wget -c "https://github.com/cosmo0/arcade-manager/releases/download/v25.1/ArcadeManager-linux-25.1.0.AppImage" -O am.AppImage
 chmod +x am.AppImage
 
 echo "Extraindo..."
